up LwIP based on "Open source LWIP for ESP_IOT_SDK_V1.4.0"

+++ include/arch/cc.h
@@ -53,9 +53,9 @@ typedef unsigned   char    u8_t;
 typedef signed     char    s8_t;
 typedef unsigned   short   u16_t;
 typedef signed     short   s16_t;
-typedef unsigned   int    u32_t;
-typedef signed     int    s32_t;
-typedef unsigned int   mem_ptr_t;
+typedef unsigned   long    u32_t;
+typedef signed     long    s32_t;
+typedef unsigned long   mem_ptr_t;
 
 #define S16_F "d"
 #define U16_F "d"
